requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-dotenv==1.4.2
django-ipware==4.0.2
django-phonenumber-field==6.0.0
django-redis==5.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
flake8==4.0.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
lxml==4.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
phonenumbers==8.12.39
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pykml==0.2.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
qrcode==7.3.1
redis==4.1.0
requests==2.26.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
